backend/requirements.txt
django>=5.0
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3
pillow>=10.3
channels>=4.0
pyjwt>=2.8
python-dotenv>=1.0
dj-database-url>=2.1
psycopg[binary]>=3.2
gunicorn>=21.2

psycopg2-binary
